refac(utils): combined read/writeClipboardText into one file
