+++ src/utils/core/index.ts
@@ -1,14 +1,13 @@
 export * from './clamp';
+export * from './clipboard';
 export * from './createShowdexEvent';
 export * from './dispatchShowdexEvent';
 export * from './getEnv';
 export * from './getExtensionId';
 export * from './getExtensionProtocol';
 export * from './getResourceUrl';
-export * from './readClipboardText';
 export * from './reverseObjectKv';
 export * from './runtimeFetch';
 export * from './safeJsonParse';
 export * from './storage';
 export * from './upsizeArray';
-export * from './writeClipboardText';
